WIP: Draft a LeiosDbConnection API
This also moves the implementations into separate modules
